Add option to randomize collect interval and optimize performance
This commit adds a new option to the ASFFreeGamesOptions class called RandomizeRecheckInterval, which is a nullable bool that indicates whether to randomize the collect interval or not. The default value is true, which means the collect interval will be randomized by default.

The commit also modifies the ASFFreeGamesPlugin class to use the RandomizeRecheckInterval option to determine the value of the RandomizeIntervalSwitch property, which is used to multiply the standard deviation of the normal distribution used to generate the random delay in the GetRandomizedTimerDelay method. If the option is false or null, then the random delay will be equal to the mean value.

The commit also optimizes the performance of the GetRandomizedTimerDelay method by checking if the standard deviation parameter is zero before calling the Random.NextGaussian method. If it is zero, then the random number will be equal to the mean parameter, and there is no need to generate a random number from a normal distribution. This can save some computation time and resources.