Fix NullReferenceException in PluginContext.get_CancellationToken()
This commit adds some checks and methods to ensure that the PluginContext is valid and initialized before using it. It also adds a cancellationToken parameter to some of the methods that use the PluginContext cancellation token. This should resolve the issue #42 (https://github.com/maxisoft/ASFFreeGames/issues/42) that was reported.